Adding more of segment tree
