+++ src/Tree/BST.h
@@ -10,6 +10,42 @@ namespace DS_VIS {
 
 namespace TREE {
 
+template<class dataType> 
+class BSTNode : public Node<dataType> {
+private: 
+    int height;
+
+    bool isBalanced;
+
+    void setHeight(int H){
+        height = H;
+    }
+
+public:
+    int getHeight()
+    {
+        return height;
+    }
+
+    bool balanced()
+    {
+        return isBalanced;
+    }
+
+    void setBalanced()
+    {
+        int left = this->getLeft() ? this->getLeft->getHeight() : 0;
+        int right = this->getRight() ? this->getRight->getHeight() : 0;
+        isBalanced = std::abs(left - right) <= 1;
+    }
+
+    BSTNode() : height{1},
+                isBalanced{false}
+    {}
+
+};
+
+
 template<class data>
 class BST final : public ITree<data> {
 private:
