+++ src/Tree/Node.h
@@ -17,10 +17,11 @@ namespace TREE{
 
 template<class data>
 class Node{
-private:
+protected:
     data val;
     std::shared_ptr<Node<data>> left;    
     std::shared_ptr<Node<data>> right;
+    
 public:
     data getVal(){
         return val;
@@ -42,40 +43,6 @@ public:
     {
         right = node;
     }
-
-};
-template<class dataType> 
-class BSTNode : public Node<dataType> {
-private: 
-    int height;
-
-    bool isBalanced;
-
-    void setHeight(int H){
-        height = H;
-    }
-
-public:
-    int getHeight()
-    {
-        return height;
-    }
-
-    bool balanced()
-    {
-        return isBalanced;
-    }
-
-    void setBalanced()
-    {
-        int left = this->getLeft() ? this->getLeft->getHeight() : 0;
-        int right = this->getRight() ? this->getRight->getHeight() : 0;
-        isBalanced = std::abs(left - right) <= 1;
-    }
-
-    BSTNode() : height{1},
-                isBalanced{false}
-    {}
 
 };
