+++ src/Tree/SegmentTree.cpp
@@ -1,0 +1,21 @@
+#include "SegmentTree.h"
+using namespace std;
+
+namespace DS_VIS 
+{
+
+namespace TREE 
+{
+
+template<class data>
+data SegmentTree<data>::query(
+    shared_ptr<SegmentTreeNode<data>> root, 
+    int l, 
+    int r)
+{
+    
+} 
+
+} // TREE
+
+} // DS_VIS
